Find all the superpermutations of the known form for n=6
Enumerate all the superpermutations that have the basic structure
of a single 3-cycle with 25 additional 2-cycles attached in a tree.

There are 42,288 of them.

I am also using this code to search for a superpermutation that
has this structure for n=7. If there is one, it will be shorter
than any that are currently known. This may be infeasible, but
it is surely worth trying.