Add code to generate supernecklaces and -bracelets
A necklace is an equivalence class of strings under rotation; a bracelet
is an equivalence class of strings under rotation and reflection.
Restricting our attention further to necklaces/bracelets without
repetitions, we can search for "supernecklaces" and "superbracelets",
which are strings that contain some representative of each
non-repeating necklace/bracelet as a substring (equivalently, strings
which, for every permutation p, contain a substring that may be rotated
and/or reflected to obtain p).

This is not a classic TSP, but it can be framed as an
Equality-Generalised Travelling Salesman Problem, which can then be
converted into an ATSP and solved by the GLKH solver,
http://webhotel4.ruc.dk/~keld/research/GLKH/.

This work was inspired by Matthew Clarke's questions on the forum:
https://groups.google.com/forum/?#!msg/superpermutators/H1_jrq_EO1c/unrp4Gl_GQAJ