Small performance boost
Small performance boost gained by starting with the cyclic successor of the previous character in each search, e.g. if a string ends with 4, for n=6, we try appending 5,6,1,2,3 rather than the previous order of 1,2,3,5,6. This has a higher chance of yielding an increase in the number of permutations visited, earlier in the search.