introduce depermutate
depermutate is a tool for finding all the permutations in a list.
It's output has been made machine-readable, as compared to demutator.
The output may be re-formatted by additional programs to become a
better visual aid.

Logic changes:
This program has been rewritten to work on streams of characters, so
that it may use less memory parsing larger lists.

The program attempts to identify the set of all elements contained in
the list it's processing before finding the permutations. This is to
eliminate the need for a separate radix, character set, or other
information besides the list itself. This assumes that all elements of
that set are already present in the list.